@@ -1,2 +1,7 @@
-package br.com.uniriotec.sagui.svgio;public class UnsuportedCurso {
+package br.com.uniriotec.sagui.svgio;
+
+public class UnsuportedCurso extends RuntimeException{
+    public UnsuportedCurso(String exception){
+        super(exception);
+    }
 }
