@@ -1,2 +1,4 @@
-package br.com.uniriotec.sagui.model;public class AlunoToModelTest {
+package br.com.uniriotec.sagui.model;
+
+public class AlunoToModelTest {
 }
