@@ -1,4 +1,4 @@
-package br.com.livrariareal.discoveryserver;
+package br.com.uniriotec.servicodescoberta;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 
 @SpringBootApplication
 @EnableEurekaServer
-public class DiscoveryServerApplication {
+public class ServicoDescobertaApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DiscoveryServerApplication.class, args);
+		SpringApplication.run(ServicoDescobertaApplication.class, args);
 	}
 
 }
